Add option to set label for hosting packages (#80)
Added possibility to set "Label" custom attribute on ordered hosting
package item. The custom attribute is used to distinguish packages in
the multi-package environment.

Configuration option "EnablePackageLabel" should be set to "true" to
allow setting a label on the package subscription and the order line.
Also, "multiPackage" option should be enabled in the billing API to
allow setting a label.

Added HTML input field to set a label for the selected package. The field
is visible on the page only if there are no domains in the cart. If
the cart has a domain, the field will not be visible nor the label will be
set on the hosting package item.

Resolves PROD-1975

ChangeLog:
* [NEW] - Added option to set a label for the ordered package.